requirements.txt
langchain
langchain-community
langchain-groq
langchain-text-splitters
langchain-huggingface
faiss-cpu
python-dotenv
pypdf
